Splitting of unittest classes into different files and storing them in the tests folder for easier testing
